+++ pyaud/__init__.py
@@ -1,14 +1,7 @@
 """Framework for writing Python package audits."""
 from . import exceptions, plugins
 from ._objects import BasePlugin
-from ._utils import files, package
+from ._utils import files
 from ._version import __version__
 
-__all__ = [
-    "BasePlugin",
-    "__version__",
-    "exceptions",
-    "files",
-    "package",
-    "plugins",
-]
+__all__ = ["BasePlugin", "__version__", "exceptions", "files", "plugins"]